+++ CHANGES.md
@@ -11,14 +11,27 @@
 
 ## develop
 
-- [UPDATE] libwebrtc m144.7559.0.1 に上げる
-  - @miosakuma @zztkm
+- [UPDATE] libwebrtc m144.7559.2.1 に上げる
+  - @t-miya
 - [UPDATE] Statistics, StatisticsEntry をドキュメント対象として公開する
   - `getStats` メソッドの返り値である `Statistics` のドキュメントを生成するため
   - @t-miya
 - [UPDATE] Configuration.simulcastRid を非推奨にする
   - 移行先は `Configuration.simulcastRequestRid`
   - @zztkm
+- [ADD] MediaChannel に `setAudioSoftMute(_:)` を追加する
+  - 音声ソフトミュート機能のシンタックスシュガー
+  - 送信ストリームの AudioTrack を取得し、MediaStream.audioEnabled を切り替える
+  - AudioTrack の有無判定を行うため、 MediaStream に `hasAudioTrack` を追加する
+  - @t-miya
+- [ADD] 音声のハードミュート有効化/無効化機能を追加する
+  - iOS 端末のマイクインジケーターを消灯させる
+  - AudioDeviceModuleWrapper クラスを追加する
+    - RTCAudioDeviceModule の pauseRecording/resumeRecording を実行するためのラッパークラス
+    - インスタンスは NativePeerChannelFactory が保持する
+  - MediaChannel に setAudioHardMute(_:) を追加する
+    - 内部で NativePeerChannelFactory 経由で AudioDeviceModuleWrapper.setAudioHardMute(_:) を呼び出す
+  - @t-miya
 - [ADD] MediaChannel に libwebrtc の統計情報を取得する `getStats` メソッドを追加する
   - @t-miya
 - [ADD] RTCAudioTrack から音声データを受け取るためのコールバックプロトコルである RTCAudioTrackSink を追加する